Balanced hunger

+++ Abstract_Class_Animal.pde
@@ -6,7 +6,7 @@ abstract class Animal {
   float maxforce;
   float oldMaxSpeed;
   float maxspeed;
-  int hunger;
+  float hunger;
 
   float size;
